services/auth-api/pyproject.toml
[tool.poetry]
name = "auth-api"
version = "0.1.0"
description = "Authentication service"
authors = ["Dan"]

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
alembic = "^1.18.0"
python-jose = "^3.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pymysql = "^1.1.2"
email-validator = "^2.3.0"
bcrypt = "<4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
